+++ lib/cuke_linter.rb
@@ -2,6 +2,7 @@ require 'cuke_modeler'
 
 require "cuke_linter/version"
 require 'cuke_linter/formatters/pretty_formatter'
+require 'cuke_linter/linters/linter'
 require 'cuke_linter/linters/example_without_name_linter'
 require 'cuke_linter/linters/feature_without_scenarios_linter'
 require 'cuke_linter/linters/outline_with_single_example_row_linter'
@@ -73,9 +74,12 @@ module CukeLinter
         # TODO: have linters lint only certain types of models
         #         linting_data.concat(linter.lint(model)) if relevant_model?(linter, model)
 
-        linted_data = linter.lint(model)
-        linted_data.each { |data_point| data_point[:linter] = linter.name }
-        linting_data.concat(linted_data)
+        result = linter.lint(model)
+
+        if result
+          result[:linter] = linter.name
+          linting_data << result
+        end
       end
     end