+++ lib/cuke_linter/linters/example_without_name_linter.rb
@@ -2,22 +2,18 @@ module CukeLinter
 
   # A linter that detects unnamed example groups
 
-  class ExampleWithoutNameLinter
+  class ExampleWithoutNameLinter < Linter
 
-    # Returns the name of the linter
-    def name
-      'ExampleWithoutNameLinter'
+    # The rule used to determine if a model has a problem
+    def rule(model)
+      return false unless model.is_a?(CukeModeler::Example)
+
+      model.name.nil? || model.name.empty?
     end
 
-    # Lints the given model and returns linting data about said model
-    def lint(model)
-      return [] unless model.is_a?(CukeModeler::Example)
-
-      if model.name.nil? || model.name.empty?
-        [{ problem: 'Example has no name', location: "#{model.get_ancestor(:feature_file).path}:#{model.source_line}" }]
-      else
-        []
-      end
+    # The message used to describe the problem that has been found
+    def message
+      'Example has no name'
     end
 
   end