+++ testing/cucumber/step_definitions/verification_steps.rb
@@ -10,7 +10,7 @@ Then(/^an error is reported$/) do |table|
   table.hashes.each do |error_record|
     expect(@results).to include({ linter:   error_record['linter'],
                                   problem:  error_record['problem'],
-                                  location: error_record['location'].sub('<path_to_file>', @model.get_ancestor(:feature_file).path) })
+                                  location: error_record['location'].sub('<path_to_file>', @model.get_ancestor(:feature_file).path).sub('<model_line_number>', @model.source_line.to_s) })
   end
 end