+++ testing/rspec/spec/unit/linters/feature_without_scenarios_linter_unit_spec.rb
@@ -44,13 +44,13 @@ RSpec.describe CukeLinter::FeatureWithoutScenariosLinter do
       end
 
       it 'records a problem' do
-        results = subject.lint(test_model_with_empty_scenarios)
+        result = subject.lint(test_model_with_empty_scenarios)
 
-        expect(results.first[:problem]).to eq('Feature has no scenarios')
+        expect(result[:problem]).to eq('Feature has no scenarios')
 
-        results = subject.lint(test_model_with_nil_scenarios)
+        result = subject.lint(test_model_with_nil_scenarios)
 
-        expect(results.first[:problem]).to eq('Feature has no scenarios')
+        expect(result[:problem]).to eq('Feature has no scenarios')
       end
 
       it 'records the location of the problem' do
@@ -61,11 +61,11 @@ RSpec.describe CukeLinter::FeatureWithoutScenariosLinter do
         model_2.tests       = []
         model_2.source_line = 3
 
-        results = subject.lint(model_1)
-        expect(results.first[:location]).to eq('path_to_file:1')
+        result = subject.lint(model_1)
+        expect(result[:location]).to eq('path_to_file:1')
 
-        results = subject.lint(model_2)
-        expect(results.first[:location]).to eq('path_to_file:3')
+        result = subject.lint(model_2)
+        expect(result[:location]).to eq('path_to_file:3')
       end
 
     end
@@ -83,7 +83,7 @@ RSpec.describe CukeLinter::FeatureWithoutScenariosLinter do
         end
 
         it 'does not record a problem' do
-          expect(subject.lint(test_model)).to eq([])
+          expect(subject.lint(test_model)).to eq(nil)
         end
 
       end
@@ -102,7 +102,7 @@ RSpec.describe CukeLinter::FeatureWithoutScenariosLinter do
         end
 
         it 'does not record a problem' do
-          expect(subject.lint(test_model)).to eq([])
+          expect(subject.lint(test_model)).to eq(nil)
         end
 
       end
@@ -111,10 +111,10 @@ RSpec.describe CukeLinter::FeatureWithoutScenariosLinter do
 
     context 'a non-feature model' do
 
-      it 'returns an empty set of results' do
-        results = subject.lint(CukeModeler::Model.new)
+      it 'returns no result' do
+        result = subject.lint(CukeModeler::Model.new)
 
-        expect(results).to eq([])
+        expect(result).to eq(nil)
       end
 
     end