+++ testing/rspec/spec/unit/linters/outline_with_single_example_row_linter_unit_spec.rb
@@ -49,9 +49,9 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         end
 
         it 'records a problem' do
-          results = subject.lint(test_model)
+          result = subject.lint(test_model)
 
-          expect(results.first[:problem]).to eq('Outline has only one example row')
+          expect(result[:problem]).to eq('Outline has only one example row')
         end
 
       end
@@ -71,9 +71,9 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         end
 
         it 'records a problem' do
-          results = subject.lint(test_model)
+          result = subject.lint(test_model)
 
-          expect(results.first[:problem]).to eq('Outline has only one example row')
+          expect(result[:problem]).to eq('Outline has only one example row')
         end
 
       end
@@ -92,11 +92,11 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         model_1.source_line = 1
         model_2.source_line = 3
 
-        results = subject.lint(model_1)
-        expect(results.first[:location]).to eq('path_to_file:1')
+        result = subject.lint(model_1)
+        expect(result[:location]).to eq('path_to_file:1')
 
-        results = subject.lint(model_2)
-        expect(results.first[:location]).to eq('path_to_file:3')
+        result = subject.lint(model_2)
+        expect(result[:location]).to eq('path_to_file:3')
       end
 
     end
@@ -117,7 +117,7 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         end
 
         it 'does not record a problem' do
-          expect(subject.lint(test_model)).to eq([])
+          expect(subject.lint(test_model)).to eq(nil)
         end
 
       end
@@ -138,7 +138,7 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         end
 
         it 'does not record a problem' do
-          expect(subject.lint(test_model)).to eq([])
+          expect(subject.lint(test_model)).to eq(nil)
         end
 
       end
@@ -159,7 +159,7 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
           end
 
           it 'does not record a problem' do
-            expect(subject.lint(test_model)).to eq([])
+            expect(subject.lint(test_model)).to eq(nil)
           end
 
         end
@@ -173,7 +173,7 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
           end
 
           it 'does not record a problem' do
-            expect(subject.lint(test_model)).to eq([])
+            expect(subject.lint(test_model)).to eq(nil)
           end
 
         end
@@ -192,7 +192,7 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         end
 
         it 'does not record a problem' do
-          expect(subject.lint(test_model)).to eq([])
+          expect(subject.lint(test_model)).to eq(nil)
         end
 
       end
@@ -211,7 +211,7 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
         end
 
         it 'does not record a problem' do
-          expect(subject.lint(test_model)).to eq([])
+          expect(subject.lint(test_model)).to eq(nil)
         end
 
       end
@@ -220,10 +220,10 @@ RSpec.describe CukeLinter::OutlineWithSingleExampleRowLinter do
 
     context 'a non-outline model' do
 
-      it 'returns an empty set of results' do
-        results = subject.lint(CukeModeler::Model.new)
+      it 'returns no result' do
+        result = subject.lint(CukeModeler::Model.new)
 
-        expect(results).to eq([])
+        expect(result).to eq(nil)
       end
 
     end